backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "cryptography>=46.0.6",
    "fastapi>=0.135.2",
    "pydantic-settings>=2.13.1",
    "uvicorn>=0.42.0",
]
